Move hashing and serializing related code to sdk
