+++ sdk/src/main/scala/sdk/hashing/Blake2Hash.scala
@@ -1,14 +1,15 @@
-package io.rhonix.rholang.parmanager.blake2
+package sdk.hashing
 
-import io.rhonix.rholang.parmanager.Constants.hashSize
 import org.bouncycastle.crypto.digests.Blake2bDigest
 
 object Blake2Hash {
+  final private val HashSize = 32
+
   @SuppressWarnings(Array("org.wartremover.warts.NonUnitStatements"))
   def hash(input: Array[Byte]): Array[Byte] = {
-    val digestFn = new Blake2bDigest(hashSize * 8)
+    val digestFn = new Blake2bDigest(HashSize * 8)
     digestFn.update(input, 0, input.length)
-    val res      = new Array[Byte](hashSize)
+    val res      = new Array[Byte](HashSize)
     digestFn.doFinal(res, 0)
     res
   }
