+++ sdk/src/main/scala/sdk/primitive/ArrayByteSyntax.scala
@@ -4,6 +4,8 @@ import java.nio.ByteBuffer
 
 trait ArrayByteSyntax {
   implicit def sdkSyntaxArrayByte(x: Array[Byte]): ArrayByteOps = new ArrayByteOps(x)
+
+  implicit val o: Ordering[Array[Byte]] = (a: Array[Byte], b: Array[Byte]) => java.util.Arrays.compare(a, b)
 }
 
 final class ArrayByteOps(private val x: Array[Byte]) extends AnyVal {
